pyproject.toml
[tool.poetry]
name = 'db-doc-builder'
version = '0.1.0'
description = 'Classify columns in Postgres schemas using DAMA categories with LLM support'
authors = ['Param Iyer paramiyer@gmail.com']

[tool.poetry.dependencies]
python = '>=3.10,<3.12'
pandas = '^2.0'
sqlalchemy = '^2.0'
openai = '^1.0'
pyyaml = '^6.0'
psycopg2-binary = "^2.9.10"
dotenv = "^0.9.9"

[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.4.3"

[build-system]
requires = ['poetry-core>=1.0.0']
build-backend = 'poetry.core.masonry.api'
